chore: renaming fsa package to client
